Fixed maximize and minimize functions
Added fixes for maximize (which toggles properly now) and minimize (does not toggle).
A non-toggling unmaximize function is ready as well, but it's obsolete now that maximize toggles.
My fixes involved xdotool, which Ben seemed shy about using for some reason. Hopefully this isn't full of inadvertent bad practices.
If I could figure out how to tell if a window is maximized I could make minimize toggle. The switch is already in place, the only code missing is the actual detecting if code is maximized.